new routes for indiviual components
